% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohesion-package.R
\docType{package}
\name{cohesion-package}
\alias{cohesion-package}
\title{cohesion: Calculates the cohesion of a cluster}
\description{
There are R packages that calculate the silhouette coefficient, but not that many that calculate the cohesion of a cluster.
}
\keyword{internal}
