% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohesion.R
\name{cohesion}
\alias{cohesion}
\title{A short description of your function}
\usage{
cohesion(cluster)
}
\arguments{
\item{x}{A matrix}
}
\value{
An integer
}
\description{
A short description of your function
}
\examples{
library(magrittr)
matrix(c(1, 2, 2, 3, 2, 2, 1, 2), ncol=2, byrow=TRUE) \%>\% cohesion()
}
\author{
Zhipei Wang
}
